create layer factories
