+++ monai/networks/blocks/convolutions.py
@@ -12,7 +12,7 @@
 import numpy as np
 import torch.nn as nn
 
-from monai.networks.utils import (get_conv_type, get_dropout_type, get_normalize_type)
+from monai.networks.layers.factories import get_conv_type, get_dropout_type, get_normalize_type
 from monai.networks.layers.convutils import same_padding
